% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-dblp.R
\docType{package}
\name{dblp}
\alias{dblp}
\alias{package-dblp}
\title{dblp}
\description{
\code{dblp} The dblp computer science bibliography at \href{https://dblp.org}{}
provides open bibliographic information on major computer science
journals and proceedings. Originally created at the University of
Trier in 1993, dblp is now operated and further developed by Schloss
Dagstuhl. This R package interfaces with the API, making data
available to use from R.
}
\keyword{package}
