% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dblp.R
\name{dblp_crawl}
\alias{dblp_crawl}
\title{Retrieve more records from dblp in batches}
\usage{
dblp_crawl(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=dblp_search]{dblp_search}}
  \describe{
    \item{\code{query}}{the query string to search for, as described in \url{https://dblp.org/faq/13501473.html}, for example q=test+search}
    \item{\code{start}}{the "f"irst hit in the numbered sequence of search results to return. In combination with the h parameter, this parameter can be used for pagination of search results}
    \item{\code{n}}{maximum number of search results ("h"its) to return. For bandwidth reasons, this number is capped at 1000}
    \item{\code{c}}{maximum number of completion terms to return. For bandwidth reasons, this number is capped at 1000}
    \item{\code{api_config}}{a configuration setting for the API including base URL etc, by default from config()}
    \item{\code{entity}}{one of "publications", "authors" or "venues" where "publications" is default}
  }}
}
\value{
results records returned from the search
}
\description{
This function retrieves batches of record (paging through results)
}
\examples{
\dontrun{
dblp_crawl(query = "chip")
}
}
