% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dblp.R
\name{dblp_search}
\alias{dblp_search}
\title{Retrieve records from dblp API}
\usage{
dblp_search(
  query = NULL,
  start = 1,
  n = 10,
  c = NULL,
  api_config = NULL,
  entity = c("publications", "authors", "venues")
)
}
\arguments{
\item{query}{the query string to search for, as described in \url{https://dblp.org/faq/13501473.html}, for example q=test+search}

\item{start}{the "f"irst hit in the numbered sequence of search results to return. In combination with the h parameter, this parameter can be used for pagination of search results}

\item{n}{maximum number of search results ("h"its) to return. For bandwidth reasons, this number is capped at 1000}

\item{c}{maximum number of completion terms to return. For bandwidth reasons, this number is capped at 1000}

\item{api_config}{a configuration setting for the API including base URL etc, by default from config()}

\item{entity}{one of "publications", "authors" or "venues" where "publications" is default}
}
\value{
results records returned from the search
}
\description{
Retrieve records from dblp API
}
\examples{
\dontrun{
dblp_search()
}


}
